Refactor toolkit acl module: rename files and functions
