+++ packages/toolkit/src/dao/acl/acl.js
@@ -3,7 +3,7 @@ import { abi as aragonAppAbi } from '@aragon/abis/os/artifacts/AragonApp'
 import { abi as kernelAbi } from '@aragon/abis/os/artifacts/Kernel'
 import { abi as repoAbi } from '@aragon/abis/os/artifacts/Repo'
 //
-import { getRecommendedGasLimit } from '../../util'
+import { getRecommendedGasLimit } from '../utils/getRecommendedGasLimit'
 
 export default function acl(web3) {
   const getACL = async repoAddr => {
