+++ packages/toolkit/src/dao/acl/index.js
@@ -1,5 +1,5 @@
 export { default as acl } from './acl'
-export * from './view'
-export * from './viewFormatter'
-export { default as grantNewVersionsPermission } from './grantNewVersionsPermission'
+export * from './getAppPermissions'
+export * from './utils/permissionsFormatter'
 export { default as getTransactionPath } from './getTransactionPath'
+export { default as grantNewVersionsPermission } from './grantNewVersionsPermission'
