+++ packages/toolkit/src/dao/acl/utils/permissionsFormatter.js
@@ -1,4 +1,4 @@
-import '../../../@types/acl/typedef'
+import '../@types/typedef'
 
 /**
  * Flattens all ACL permissions from an acl object into a single array
